+++ app/Http/Controllers/TeamController.php
@@ -1,12 +1,16 @@
 <?php
 
+
 namespace App\Http\Controllers;
+
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 use Illuminate\Http\Request;
 use App\Models\Team;
 
 class TeamController extends Controller
 {
+    use AuthorizesRequests;
     public function index()
     {
         $teams = Team::all();
@@ -14,10 +18,12 @@ class TeamController extends Controller
     }
     public function create()
     {
+        $this->authorize('create', Team::class);
         return view('teams.create');
     }
     public function store(Request $request)
     {
+        $this->authorize('create', Team::class);
         $request->validate(['name' => 'required']);
         Team::create($request->all());
         return redirect()->route('teams.index');
@@ -28,16 +34,19 @@ class TeamController extends Controller
     }
     public function edit(Team $team)
     {
+        $this->authorize('update', $team);
         return view('teams.edit', compact('team'));
     }
     public function update(Request $request, Team $team)
     {
+        $this->authorize('update', $team);
         $request->validate(['name' => 'required']);
         $team->update($request->all());
         return redirect()->route('teams.index');
     }
     public function destroy(Team $team)
     {
+        $this->authorize('delete', $team);
         $team->delete();
         return redirect()->route('teams.index');
     }
